[add] ask user before deleting elements

+++ source/YumlFrontEnd/Command/Classifier/DeleteClassifierCommand.cs
@@ -9,17 +9,22 @@ namespace Yuml.Command
     public class DeleteClassifierCommand : DomainObjectBaseCommand<Classifier>, IDeleteCommand
     {
         private readonly DeletionService _deletionService;
+        private readonly IAskUserBeforeDeletionService _askUserBeforeDeletion;
 
         public DeleteClassifierCommand(
             Classifier classifier,
-            DeletionService deletionService) : base(classifier) // message handling is done by service
+            DeletionService deletionService,
+            IAskUserBeforeDeletionService askUserBeforeDeletion) : base(classifier) // message handling is done by service
         {
             _deletionService = deletionService;
+            _askUserBeforeDeletion = askUserBeforeDeletion;
         }
 
         public void DeleteItem()
         {
-            _deletionService.DeleteClassifier(_domainObject);
+            if (_askUserBeforeDeletion == null ||
+                _askUserBeforeDeletion.ShouldDomainObjectBeDeleted(Strings.AskBeforeDeletingClassifier))
+                _deletionService.DeleteClassifier(_domainObject);
         }
     }
 }
